+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Revision history for patch
 
+## 0.1.0.0
+
+* Rewrite `PatchMapWithMove` in terms of `PatchMapWithPatchingMove`.
+  Care is taken to make this as little of a breaking change as possible.
+  In particular, `PatchMapWithMove` is a newtype of `PatchMapWithPatchingMove` as is the `NodeInfo` of `PatchMapWithPatchingMove`'s `NodeInfo`.
+
 ## 0.0.3.0
 
 * Create `PatchMapWithPatchingMove` variant which supports moves with a patch.